backend/requirements.txt
# CareerBoost Backend Requirements

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
groq==0.11.0
PyPDF2==3.0.1
python-dotenv==1.0.1

# Optional dependencies for production
# pymongo==4.10.1  # If using MongoDB
# sqlalchemy==2.0.36  # If using SQL database
# python-jose[cryptography]==3.3.0  # For JWT authentication
# passlib[bcrypt]==1.7.4  # For password hashing
# python-docx==1.1.2  # For Word document parsing
